requirements.txt
python-telegram-bot
binance-connector
pyjwt[crypto]